+++ src/containers/either.js
@@ -41,6 +41,42 @@ const
     };
 
     return new Right(value);
+  },
+
+  // left :: a -> Left(a)
+  // Implements the "Left of a" pattern.
+  left = value => {
+    let
+      // Contained value is kept in a JavaScript closure to avoid direct access
+      // to it from outside.
+      _value = null;
+
+    const
+      // Prototype for the container object exposed to outside.
+      Left = newValue => {
+        if('undefined' !== typeof newValue) {
+          _value = deepFreeze(newValue);
+        }
+      };
+
+    // map :: _ -> Left(a)
+    // Note: not using arrow function as `this` must bind to the inner scope.
+    Left.prototype.map = function() {
+      return this;
+    };
+
+    // unsafeMap :: (_) -> Left(a)
+    // Note: not using arrow function as `this` must bind to the inner scope.
+    Left.prototype.unsafeMap = function(f) {
+      if('function' === typeof f) {
+        // When given `f` is callable, call it and discard the return value.
+        f(_value);
+      }
+
+      return this;
+    };
+
+    return new Left(value);
   };
 
-export { right };
+export { right, left };